Fix für #208 bei schrittweiser Indexierung ist Index nicht vollständig

+++ lib/search_it.php
@@ -416,7 +416,7 @@ class search_it
         }
 
         // delete from index
-            $delete->delete();
+        if ( $_start === false || $_start == 0 ) { $delete->delete(); }
 
 
         // NEW
